refactor(schema): modularize reference model
Split the reference model out of the 3377-line facade.

Move constructors, serde/schema impls, accessors, discriminator tests,
normalization tests, numbering tests, InputReference, and class types
into focused sibling modules.

Keep citum_schema::reference re-exports and the flat class-discriminator
wire format stable. Centralize pure class dispatch for metadata and
shared field access while leaving class-specific behavior explicit.

Update the modularization bean to reflect the completed module boundary.
